server/requirements.txt
Flask==1.1.2
Flask-Limiter==1.3.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
PyJWT==1.7.1
python-dotenv==0.14.0
psycopg2-binary==2.8.5